+++ src/devpilot_core/application/__init__.py
@@ -60,6 +60,7 @@ from .settings_service import SettingsApplicationService
 from .services import ApplicationService
 from .pre_code_wizard_service import PreCodeWizardApplicationService
 from .roadmap_workbench_service import RoadmapWorkbenchApplicationService
+from .backlog_workbench_service import BacklogWorkbenchApplicationService
 from .validation_service import ValidationApplicationService
 from .workspace_service import WorkspaceApplicationService
 from .workspace_documents_service import WorkspaceDocumentsApplicationService
@@ -124,6 +125,7 @@ __all__ = [
     "ApplicationService",
     "PreCodeWizardApplicationService",
     "RoadmapWorkbenchApplicationService",
+    "BacklogWorkbenchApplicationService",
     "EvaluationApplicationService",
     "HistoryApplicationService",
     "InterfaceRouteContract",